feat: restore ControlEntryDetailsScreen
