+++ packages/apps/quality/src/screens/ControlEntryDetailsScreen.js
@@ -32,7 +32,6 @@ const ControlEntryDetailsScreen = ({route}) => {
 
   const dispatch = useDispatch();
   const I18n = useTranslator();
-  const Colors = useThemeColor();
 
   const {
     controlEntrySampleList,
